requirements.txt
streamlit
langchain-groq
langchain-community
langchain-chroma
langchain-huggingface
sentence-transformers
chromadb
pydantic
python-dotenv